changed int to double

+++ src/main.cpp
@@ -26,7 +26,7 @@ int main (int argc, char** argv)
     int number_qbs;
     int i;
     int max_y;
-    int max_x;
+    double max_x;
     std::vector<Quarterback*> qbs;
     std::multimap<double, Quarterback*> sorted_qbs;
     Quarterback *tmp;
